chore(test framework): refactor config models to move helper methods out to utilities (#2470)
* chore:refactor config models


* trigger PR


---------
